// sourcecode/BTrak.Packages/snova-app-builder-creation-components/app-builder-module-components/app-builder-components/src/lib/builder/components/genericform/models/custom-application-input.model.ts
export class CustomApplicationModel {
    customApplicationId: string;
    customApplicationName: string;
    formTypeId: string;
    formTypeName: string;
    formId: string;
    formName: string;
    isArchived: boolean;
    createdDateTime: Date;
    createdByUserId: string;
    originalId: string;
    timeStamp: any;
    asAtInactiveDateTime: string;
    totalCount: number;
    roleIds: string;
    formJson: string;
    publicMessage: string;
    domainName: string;
    selectedKeyIds: string;
    selectedPrivateKeyIds: string;
    selectedTagKeyIds: string;
    selectedEnableTrendsKeys: string;
    isPublished: boolean;
    description: string;
    tag: string;
    moduleIds: string;
    selectedForms: any[];
    fields: any;
    isPdfRequired : boolean;
    isApproveNeeded:any;
    userList:string;
    isRedirectToEmails: boolean;
    workflowIds : string;
    allowAnnonymous:boolean;
    toEmails : string;
    toRoleIds : string;
    message : string;
    subject : string;
    isRecordLevelPermissionEnabled: boolean;
    recordLevelPermissionFieldName: string;
    conditionalEnum: any;
    approveSubject: string;
    approveMessage: string;
    conditionsJson: any;
    filters: any;
    stageScenariosJson: any;
    stagesScenarios : any;
}